% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geq.R
\name{geq}
\alias{geq}
\title{Title General estimation equations for the covariate coefficients estimation}
\usage{
geq(Time, Status, X, beta, id, eps, corstr, itermax)
}
\arguments{
\item{Time}{right censored data which is the follow up time.}

\item{Status}{right censored data which is the follow up time.}

\item{X}{a matrix of covariates that that may have effect on the failure times.}

\item{beta}{the initial value of the covariate coefficients.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{eps}{tolerance for convergence. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}

\item{itermax}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{itermax} iterations and
the estimates will be based on the last iteration.}
}
\description{
Estimation the covariate coefficients using the GEE approach.
}
