% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqbeta.R
\name{getlambda}
\alias{getlambda}
\title{Title An auxiliary function for solving lambda given beta}
\usage{
getlambda(lam, dRbet)
}
\arguments{
\item{lam}{unknown parameter corresponding to the largrane multiplier.}

\item{dRbet}{a matrix needed in this function.}
}
\description{
An auxiliary function for solving lambda given beta
}
