% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqbeta.R
\name{gqbeta}
\alias{gqbeta}
\title{Title Estimate the covariate coefficients using the GEE or QIF approach}
\usage{
gqbeta(Time, Status, X, stad, Ibeta, id, itermax, eps, method, corstr)
}
\arguments{
\item{Time}{right censored data which is the follow up time.}

\item{Status}{the censoring indicator, normally 0 = event of interest happens, and 0 = censoring}

\item{X}{a matrix of covariates that may have effect on the failure times.}

\item{stad}{if it is TRUE, all the covariates in the \code{formula} are standardized.}

\item{Ibeta}{the initial value of the covariate coefficients.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{itermax}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{itermax} iterations and
the estimates will be based on the last iteration.}

\item{eps}{tolerance for convergence. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{method}{using GEE or QIF method to estimate the covariate coefficients.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}
}
\description{
Estimate the covariate coefficients using the GEE or QIF approach. For the GEE approach, an estimation of correlation coefficient
in the working correlation structure is given.
}
