% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marptc.R
\name{print.marptc}
\alias{print.marptc}
\title{Title Print marptc object}
\usage{
\method{print}{marptc}(x, ...)
}
\arguments{
\item{x}{an object of \code{marptc}.}

\item{...}{further arguments to be added in the \code{print.marptc} function.}
}
\description{
Title Print marptc object
}
