% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qifptc.R
\name{print.qifptc}
\alias{print.qifptc}
\title{Title Print qifptc object}
\usage{
\method{print}{qifptc}(x, ...)
}
\arguments{
\item{x}{an object of \code{qifptc}.}

\item{...}{further arguments to be added in the \code{print.qifptc} function.}
}
\description{
Title Print qifptc object
}
