% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qifptc.R
\name{qifptc}
\alias{qifptc}
\title{Title  Fit the marginal semiparametric promotion time cure model under clustered survival time}
\usage{
qifptc(
  formula,
  data,
  id,
  Var = TRUE,
  Ibeta = NULL,
  stad = TRUE,
  boots = FALSE,
  nboot = 100,
  method = "GEE",
  corstr = "independence",
  itermax = 500,
  eps = 1e-06
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}. The \code{response} is a \code{Surv} object with right censoring.
The expression to the right of the "~" specifies the effect of covariates on the failure time.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula}.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{Var}{if it is TRUE, the program returns Std.Error. By default, \code{Var = TRUE}.}

\item{Ibeta}{the initial value of the covariate coefficients.}

\item{stad}{if it is TRUE, all the covariates in the \code{formula} are standardized. By default, \code{stdz = TRUE}.}

\item{boots}{if it is TRUE, the program returns Std.Error by the bootstrap method. By default, \code{boots = FALSE}}

\item{nboot}{the number of bootstrap samples. The default is \code{nboot = 100}.}

\item{method}{using GEE or QIF method to estimate the covariate coefficients.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}

\item{itermax}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{esmax} iterations and
the estimates will be based on the last iteration. The default \code{itermax = 500}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}
}
\value{
An object of class \code{qifptc} is returned. It can be examined by \code{print.qifptc()}.
}
\description{
Fit the marginal promotion time cure (PTC) model based on the general estimation equations (GEE) and the quadratic inference functions (QIF).
We consider three common correlation structures (\code{independence}, \code{exchangeable} and \code{AR1}) in this funciton.
The variances can be consistently estimated by a sandwich variance estimator. We also present a bootstrap procedure for variance estimation.
}
