% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vargee.R
\name{vargee}
\alias{vargee}
\title{Title  Estimate Variance of coefficients estimation from the GEE approach with a sandwich formula}
\usage{
vargee(Time, Status, X, id, beta, corstr)
}
\arguments{
\item{Time}{right censored data which is the follow up time.}

\item{Status}{the censoring indicator, normally 0 = event of interest happens, and 0 = censoring}

\item{X}{a matrix of covariates that may have effect on the failure times.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{beta}{covariate coefficients estimation from GEE method.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}
}
\description{
Calculate the variance estimates using the sandwich formula.
}
